Make fully functional
Following the last update, the program did not really function well,
particularly in the area of comparing inputed strings with those available
on record. Several changes were made across different files. The function
initially used to validate file opening was commented out as well as
variables for testing this.

The output on the console is still not satifactory in terms of aesthetics
and this will be addressed later.